+++ azure/datalake/store/__init__.py
@@ -6,6 +6,7 @@
 # license information.
 # --------------------------------------------------------------------------
 
+__version__ = "0.0.17"
 
 from .core import AzureDLFileSystem
 from .multithread import ADLDownloader
@@ -16,4 +17,3 @@ import logging
 from logging import NullHandler
 
 logging.getLogger(__name__).addHandler(NullHandler())
-__version__ = "0.0.14"