+++ azure/datalake/store/core.py
@@ -19,6 +19,8 @@ import io
 import logging
 import sys
 import time
+import uuid
+
 
 # local imports
 from .exceptions import DatalakeBadOffsetException
@@ -667,6 +669,9 @@ class AzureDLFile(object):
         self.buffer = io.BytesIO()
         self.blocksize = blocksize
         self.first_write = True
+        uniqueid = str(uuid.uuid4())
+        self.filesessionid = uniqueid
+        self.leaseid = uniqueid
         
         # always invalidate the cache when checking for existence of a file
         # that may be created or written to (for the first time).
@@ -758,11 +763,11 @@ class AzureDLFile(object):
             self.start = start
             self.end = min(end + self.blocksize, self.size)
             response = _fetch_range_with_retry(
-                self.azure.azure, self.path.as_posix(), start, self.end)
+                self.azure.azure, self.path.as_posix(), start, self.end, filesessionid=self.filesessionid)
             self.cache = getattr(response, 'content', response)
         if start < self.start:
             response = _fetch_range_with_retry(
-                self.azure.azure, self.path.as_posix(), start, self.start)
+                self.azure.azure, self.path.as_posix(), start, self.start, filesessionid=self.filesessionid)
             new = getattr(response, 'content', response)
             self.start = start
             self.cache = new + self.cache
@@ -771,7 +776,7 @@ class AzureDLFile(object):
                 return
             newend = min(self.size, end + self.blocksize)
             response = _fetch_range_with_retry(
-                self.azure.azure, self.path.as_posix(), self.end, newend)
+                self.azure.azure, self.path.as_posix(), self.end, newend, filesessionid=self.filesessionid)
             new = getattr(response, 'content', response)
             self.end = newend
             self.cache = self.cache + new
@@ -820,13 +825,14 @@ class AzureDLFile(object):
             raise ValueError('File not in write mode')
         if self.closed:
             raise ValueError('I/O operation on closed file.')
+        
         out = self.buffer.write(ensure_writable(data))
         self.loc += out
-        if self.buffer.tell() >= self.blocksize:
-            self.flush()
+        self.flush(syncFlag='DATA')
         return out
+        
 
-    def flush(self, force=False):
+    def flush(self, syncFlag='METADATA', force=False):
         """
         Write buffered data to ADL.
 
@@ -841,7 +847,11 @@ class AzureDLFile(object):
         """
         if not self.writable() or self.closed:
             return
-
+        
+        if not (syncFlag == 'METADATA' or syncFlag == 'DATA' or syncFlag == 'CLOSE'):
+            raise ValueError('syncFlag must be one of these: METADAT, DATA or CLOSE')
+        
+        
         if self.buffer.tell() == 0:
             if force and self.first_write:
                 _put_data_with_retry(
@@ -850,66 +860,82 @@ class AzureDLFile(object):
                     path=self.path.as_posix(),
                     data=None,
                     overwrite='true',
-                    write='true')
+                    write='true',
+                    syncFlag=syncFlag,
+                    leaseid=self.leaseid,
+                    filesessionid=self.filesessionid)
                 self.first_write = False
             return
 
         self.buffer.seek(0)
         data = self.buffer.read()
-
-        if self.delimiter:
-            while len(data) >= self.blocksize:
+        
+        syncFlagLocal = 'DATA'
+        while len(data) > self.blocksize:
+            if self.delimiter:
                 place = data[:self.blocksize].rfind(self.delimiter)
-                if place < 0:
-                    # not found - write whole block
-                    limit = self.blocksize
-                else:
-                    limit = place + len(self.delimiter)
-                if self.first_write:
-                    _put_data_with_retry(
-                        self.azure.azure,
-                        'CREATE',
-                        path=self.path.as_posix(),
-                        data=data[:limit],
-                        overwrite='true',
-                        write='true')
-                    self.first_write = False
-                else:
-                    _put_data_with_retry(
-                        self.azure.azure,
-                        'APPEND',
-                        path=self.path.as_posix(),
-                        data=data[:limit],
-                        append='true')
-                logger.debug('Wrote %d bytes to %s' % (limit, self))
-                data = data[limit:]
-            self.buffer = io.BytesIO(data)
-            self.buffer.seek(0, 2)
-
-        if not self.delimiter or force:
+            else:
+                place = -1
+            if place < 0:
+                # not found - write whole block
+                limit = self.blocksize
+            else:
+                limit = place + len(self.delimiter)
+            if self.first_write:
+                _put_data_with_retry(
+                    self.azure.azure,
+                    'CREATE',
+                    path=self.path.as_posix(),
+                    data=data[:limit],
+                    overwrite='true',
+                    write='true',
+                    syncFlag=syncFlagLocal,
+                    leaseid=self.leaseid,
+                    filesessionid=self.filesessionid)
+                self.first_write = False
+            else:
+                _put_data_with_retry(
+                    self.azure.azure,
+                    'APPEND',
+                    path=self.path.as_posix(),
+                    data=data[:limit],
+                    append='true',
+                    syncFlag=syncFlagLocal,
+                    leaseid=self.leaseid,
+                    filesessionid=self.filesessionid)
+            logger.debug('Wrote %d bytes to %s' % (limit, self))
+            data = data[limit:]
+            
+                
+        self.buffer = io.BytesIO(data)
+        self.buffer.seek(0, 2)
+            
+        if force:
             zero_offset = self.tell() - len(data)
-            offsets = range(0, len(data), self.blocksize)
-            for o in offsets:
-                offset = zero_offset + o
-                d2 = data[o:o+self.blocksize]
-                if self.first_write:
-                    _put_data_with_retry(
-                        self.azure.azure,
-                        'CREATE',
-                        path=self.path.as_posix(),
-                        data=d2,
-                        overwrite='true',
-                        write='true')
-                    self.first_write = False
-                else:
-                    _put_data_with_retry(
-                        self.azure.azure,
-                        'APPEND',
-                        path=self.path.as_posix(),
-                        data=d2,
-                        offset=offset,
-                        append='true')
-                logger.debug('Wrote %d bytes to %s' % (len(d2), self))
+            if self.first_write:
+                _put_data_with_retry(
+                    self.azure.azure,
+                    'CREATE',
+                    path=self.path.as_posix(),
+                    data=data,
+                    overwrite='true',
+                    write='true',
+                    syncFlag=syncFlag,
+                    leaseid=self.leaseid,
+                    filesessionid=self.filesessionid)
+                self.first_write = False
+            else:
+                _put_data_with_retry(
+                    self.azure.azure,
+                    'APPEND',
+                    path=self.path.as_posix(),
+                    data=data,
+                    offset=zero_offset,
+                    append='true',
+                    syncFlag=syncFlag,
+                    leaseid=self.leaseid,
+                    filesessionid=self.filesessionid)
+            logger.debug('Wrote %d bytes to %s' % (len(data), self))
             self.buffer = io.BytesIO()
 
     def close(self):
@@ -917,10 +943,11 @@ class AzureDLFile(object):
 
         If in write mode, causes flush of any unwritten data.
         """
+        logger.info("closing stream")
         if self.closed:
             return
         if self.writable():
-            self.flush(force=True)
+            self.flush(syncFlag='CLOSE', force=True)
             self.azure.invalidate_cache(self.path.as_posix())
         self.closed = True
 
@@ -948,20 +975,20 @@ class AzureDLFile(object):
         self.close()
 
 
-def _fetch_range(rest, path, start, end, stream=False):
+def _fetch_range(rest, path, start, end, stream=False, **kwargs):
     logger.debug('Fetch: %s, %s-%s', path, start, end)
     # if the caller gives a bad start/end combination, OPEN will throw and
     # this call will bubble it up
     return rest.call(
-        'OPEN', path, offset=start, length=end-start, read='true', stream=stream)
+        'OPEN', path, offset=start, length=end-start, read='true', stream=stream, **kwargs)
 
 
 def _fetch_range_with_retry(rest, path, start, end, stream=False, retries=10,
-                            delay=0.01, backoff=3):
+                            delay=0.01, backoff=3, **kwargs):
     err = None
     for i in range(retries):
         try:
-            return _fetch_range(rest, path, start, end, stream=False)
+            return _fetch_range(rest, path, start, end, stream=False, **kwargs)
         except Exception as e:
             err = e
             logger.debug('Exception %s on ADL download on attempt: %s, retrying in %s seconds',