+++ tests/conftest.py
@@ -16,7 +16,7 @@ from tests.testing import working_dir
 @pytest.fixture(scope="session", autouse=True)
 def setup_env(request):
     home = working_dir()
-    fs = AzureDLFileSystem(store=settings.STORE_NAME, token=settings.TOKEN)
+    fs = AzureDLFileSystem(store_name=settings.STORE_NAME, token=settings.TOKEN)
     if settings.RECORD_MODE != 'none':
         if not fs.exists(home):
             fs.mkdir(home)