+++ tests/fake_settings.py
@@ -5,8 +5,9 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
+import os
 
-STORE_NAME = 'fakestore'
-TENANT_ID = 'faketenant'
-SUBSCRIPTION_ID = 'fakesubscription'
-RESOURCE_GROUP_NAME = 'fakeresourcegroup'
+STORE_NAME = os.environ['azure_data_lake_store_name']
+TENANT_ID = os.environ['azure_tenant_id']
+SUBSCRIPTION_ID = os.environ['azure_subscription_id']
+RESOURCE_GROUP_NAME = os.environ['azure_resource_group_name']