+++ tests/test_transfer.py
@@ -16,7 +16,6 @@ from tests.testing import azure, posix
 
 def test_shutdown(azure):
     def transfer(adlfs, src, dst, offset, size, blocksize, buffersize, retries=5, shutdown_event=None):
-        time.sleep(1.0)
         while shutdown_event and not shutdown_event.is_set():
             time.sleep(0.1)
         return size, None
@@ -25,7 +24,6 @@ def test_shutdown(azure):
                                chunked=False)
     client.submit('foo', 'bar', 16)
     client.run(monitor=False)
-    time.sleep(0.1)
     client.shutdown()
 
     assert client.progress[0].state == 'finished'
@@ -33,7 +31,6 @@ def test_shutdown(azure):
 
 def test_submit_and_run(azure):
     def transfer(adlfs, src, dst, offset, size, blocksize, buffersize, shutdown_event=None):
-        time.sleep(0.1)
         return size, None
 
     client = ADLTransferClient(azure, transfer=transfer, chunksize=8,
@@ -63,9 +60,54 @@ def test_submit_and_run(azure):
                                                for chunk in f.chunks])
 
 
+def test_update_progress(azure):
+    """
+    Upload a 32 bytes file in chunks of 8 and test that the progress is incrementally
+    updated.
+    """
+    calls = []
+
+    def recording_callback(progress, total):
+        calls.append((progress, total))
+
+    def transfer(adlfs, src, dst, offset, size, blocksize, buffersize, shutdown_event=None):
+        return size, None
+
+    client = ADLTransferClient(azure, transfer=transfer, chunksize=8,
+                               chunked=True, progress_callback=recording_callback)
+
+    client.submit('foo', AzureDLPath('bar'), 32)
+    client.run()
+
+    assert calls == [(8, 32), (16, 32), (24, 32), (32, 32)]
+
+
+def test_merge(azure):
+
+    calls = []
+
+    def merge(adlfs, outfile, files, shutdown_event=None, overwrite=False):
+        calls.append(files)
+
+    def transfer(adlfs, src, dst, offset, size, blocksize, buffersize, shutdown_event=None):
+        return size, None
+
+    class XLoaderMock(object):
+        _overwrite = False
+
+    file_size = 32
+    chunk_size = 8
+    client = ADLTransferClient(azure, parent=XLoaderMock(), transfer=transfer, merge=merge,
+                               chunksize=chunk_size, chunked=True)
+
+    client.submit('foo', AzureDLPath('bar'), file_size)
+    client.run()
+
+    assert len(calls[0]) == file_size / chunk_size
+
+
 def test_temporary_path(azure):
     def transfer(adlfs, src, dst, offset, size, blocksize, buffersize):
-        time.sleep(0.1)
         return size, None
 
     client = ADLTransferClient(azure, transfer=transfer, chunksize=8,